feat(user_secrets): add test to validate the StatefulSet creation
This test ensures that the patch added for the user secrets part
ends up creating the proper resources in the final StatefulSet.

Caveat: it's currently not possible to fully validate the whole
chain as is requires both the data-services and secrets-storage
service to be running. This will be donne at a later stage.