fix: Mount /dev/shm to the correct container.
Currently container #1 is `oauth-proxy` while JupyterServer,
is #0.

Additionally use half of the memory requested by the pod
instead of half of the storage.

Since we rely on memory there is no need to tie the increased
/dev/shm to requesting storage in the session.

NOTE: This system of patching the manifest is very brittle and
the bug being fixed by this commit might appear again as soon
as the containers sorting changes. We should think about a different
way of interacting with the manifests of the user-sessions.