fix: image checks for some oci images failed
This occurs because some images (most likely when they support multiple
architectures) have an index instead of just a manifest. An OCI index
refers to a set of images. But when we were checking for images we were
only checking for the manifest and did not have the option to support an
index content type.